requirements.txt
kafka-python-ng
requests
python-dotenv
transformers
torch
schedule

